+++ R/plot.R
@@ -7,10 +7,10 @@
 ##' @param prob A logical variable indicating block probabilities are printed on the plot. Default is FALSE.
 ##' @param digits integer indicating the number of decimal places for probability
 ##' @param prob.cex A numerical value of `cex` of probabilites. 
-##' @param prob.col color of probabilites.
+##' @param prob.col color of probabilities on each bin.
 ##' @param ... other arguments to pass to [stats::heatmap()]. See details.
 ##' @details 
-##' ... includes various graphical parameters passes to [stats::heatmap()], then [graphics::image()]. 
+##' ... includes various [`graphical parameters`] passes to [stats::heatmap()], then [graphics::image()]. 
 ##' @returns 
 ##' a heatmap of `p_mat` orderd by `idx_order` in ``nethist`` object.
 ##' @examples
@@ -24,13 +24,20 @@
 ##' #with user-specified order
 ##' idx<- unique(hist_A$cluster) 
 ##' plot(hist_A, idx_order = idx)
+##' 
+##' #User-speicifc bin color pallete (see [graphical parameters])
+##' plot(hist_polblog,  idx_order = ind, col = colorRampPalette(colors=c("#FFFFFF","#000000"))(50))
+##' 
+##' #Users can print p_mat on the plot using user-specific colors
+##' plot(hist_polblog,  idx_order = ind, prob= TRUE, prob.col = "blue",
+##'       col = colorRampPalette(colors=c("#FFFFFF","#000000"))(50))
 ##' }
 ##' @importFrom stats heatmap
 ##' @exportS3Method 
 ##' @export
 plot.nethist <- function(x, idx_order = 1:max(x$cluster), 
                          prob = FALSE, digits = 2,
-                         prob.cex =  0.1 + 0.5/log10(k),
+                         prob.cex =  0.1 + 0.5/log10(max(x$cluster)),
                          prob.col = "black",
                          ...){
   k<-max(x$cluster)